requirements.txt
django-filter==2.0.0
djangorestframework==3.8.2
django-cors-headers==2.4.0
django-rest-auth==0.9.3
django-allauth==0.38
pytest==3.4.1
pytest-cov==2.5.1
pytest-django==3.1.2
whitenoise==3.2.2
django-spa
psycopg2
channels_redis
Django<=3.0
gunicorn==19.9.0
djangorestframework-jwt==1.11.0
requests==2.19.1
markdown==2.6.11
